Fixing Restaurant for 2nd time

+++ Food-Delivery-BackEndd/src/main/java/com/example/fooddeliveryy/Repositories/logInRepo.java
@@ -6,9 +6,9 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface logInRepo extends JpaRepository<User, Integer> {
+public interface logInRepo extends JpaRepository<User, Long> {
     Optional<User> findByUserEmail(String email);
-    Optional<User> findByUserId(Long id);
+    Optional<User> findByUserId(long id);
     boolean existsByUserEmail(String email);
 
 }
